Refactoring: Change package of some upnp classes
