+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/IndexProc.java
@@ -32,7 +32,6 @@ import com.tesshu.jpsonic.domain.MusicFolderContent;
 import com.tesshu.jpsonic.domain.MusicIndex;
 import com.tesshu.jpsonic.service.MediaFileService;
 import com.tesshu.jpsonic.service.MusicIndexService;
-import com.tesshu.jpsonic.service.upnp.ProcId;
 import com.tesshu.jpsonic.service.upnp.processor.composite.IndexOrSong;
 import org.jupnp.support.model.DIDLContent;
 import org.jupnp.support.model.container.Container;
