+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/RecentAlbumId3Proc.java
@@ -25,7 +25,6 @@ import java.util.concurrent.ExecutionException;
 import com.tesshu.jpsonic.dao.AlbumDao;
 import com.tesshu.jpsonic.domain.Album;
 import com.tesshu.jpsonic.service.MediaFileService;
-import com.tesshu.jpsonic.service.upnp.ProcId;
 import org.jupnp.support.model.BrowseResult;
 import org.jupnp.support.model.DIDLContent;
 import org.springframework.stereotype.Service;
