+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/RecentAlbumProc.java
@@ -24,7 +24,6 @@ import java.util.concurrent.ExecutionException;
 
 import com.tesshu.jpsonic.domain.MediaFile;
 import com.tesshu.jpsonic.service.MediaFileService;
-import com.tesshu.jpsonic.service.upnp.ProcId;
 import org.jupnp.support.model.BrowseResult;
 import org.jupnp.support.model.DIDLContent;
 import org.springframework.stereotype.Service;
