+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/SongByGenreProc.java
@@ -26,7 +26,6 @@ import java.util.List;
 import com.tesshu.jpsonic.domain.Genre;
 import com.tesshu.jpsonic.domain.MediaFile;
 import com.tesshu.jpsonic.service.SearchService;
-import com.tesshu.jpsonic.service.upnp.ProcId;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.jupnp.support.model.DIDLContent;
 import org.jupnp.support.model.container.Container;
